# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Atlas API",
    "version": "1.0.0",
    "description": "Contract for the Atlas platform API. This document is the source of truth; implementation follows it (blueprint/14 - API.md, ADR-001 Contract First). Errors follow RFC 7807; collections use cursor pagination."
  },
  "servers": [{ "url": "/", "description": "Relative to the API host" }],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/progress/measurements": {
      "get": {
        "tags": ["Progress"],
        "summary": "List the user's measurement history (cursor-paginated)",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of measurements (newest first)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeasurementPage" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Progress"],
        "summary": "Record (or replace) a measurement for a date",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RecordMeasurementRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The recorded measurement",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeasurementView" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/v1/progress/measurements/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" }
        }
      ],
      "delete": {
        "tags": ["Progress"],
        "summary": "Delete a measurement",
        "responses": {
          "204": { "description": "Measurement deleted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/progress/summary": {
      "get": {
        "tags": ["Progress"],
        "summary": "Derived records, stats, and the weight chart series",
        "responses": {
          "200": {
            "description": "Progress summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProgressSummary" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or failed",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
...
      "ValidationError": {
        "description": "Request failed validation",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found (or not owned by the caller)",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      }
    },
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "category", "traceId"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" },
          "category": {
            "type": "string",
            "enum": [
              "validation",
              "authentication",
              "authorization",
              "business_rule",
              "conflict",
              "not_found",
              "rate_limit",
              "infrastructure",
              "unexpected"
            ]
          },
          "traceId": { "type": "string" },
          "issues": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationIssue" }
          }
        }
      },
      "ValidationIssue": {
        "type": "object",
        "required": ["field", "message", "code"],
        "properties": {
          "field": { "type": "string" },
          "message": { "type": "string" },
          "code": { "type": "string" }
        }
      },
...
      "BodyMeasurements": {
        "type": "object",
        "properties": {
          "waist": { "type": ["number", "null"] },
          "chest": { "type": ["number", "null"] },
          "hips": { "type": ["number", "null"] },
          "arm": { "type": ["number", "null"] },
          "thigh": { "type": ["number", "null"] },
          "calf": { "type": ["number", "null"] }
        }
      },
      "MeasurementView": {
        "type": "object",
        "required": ["id", "recordedOn", "weightKg", "measurements", "note", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "recordedOn": { "type": "string", "format": "date" },
          "weightKg": { "type": ["number", "null"] },
          "measurements": { "$ref": "#/components/schemas/BodyMeasurements" },
          "note": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "MeasurementPage": {
        "type": "object",
        "required": ["items", "nextCursor", "previousCursor", "hasNext"],
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MeasurementView" }
          },
          "nextCursor": { "type": ["string", "null"] },
          "previousCursor": { "type": ["string", "null"] },
          "hasNext": { "type": "boolean" }
        }
      },
      "RecordMeasurementRequest": {
        "type": "object",
        "properties": {
          "recordedOn": { "type": "string", "format": "date" },
          "weightKg": { "type": ["number", "null"] },
          "measurements": { "$ref": "#/components/schemas/BodyMeasurements" },
          "note": { "type": ["string", "null"], "maxLength": 280 }
        }
      },
      "WeightPoint": {
        "type": "object",
        "required": ["recordedOn", "weightKg"],
        "properties": {
          "recordedOn": { "type": "string", "format": "date" },
          "weightKg": { "type": "number" }
        }
      },
      "ProgressSummary": {
        "type": "object",
        "required": [
          "entryCount",
          "latestWeightKg",
          "lowestWeightKg",
          "highestWeightKg",
          "totalChangeKg",
          "weightSeries"
        ],
        "properties": {
          "entryCount": { "type": "integer" },
          "latestWeightKg": { "type": ["number", "null"] },
          "lowestWeightKg": { "type": ["number", "null"] },
          "highestWeightKg": { "type": ["number", "null"] },
          "totalChangeKg": { "type": ["number", "null"] },
          "weightSeries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/WeightPoint" }
          }
        }
      }
    }
  }
}